requirements.txt
allure-pytest==2.12.0
Appium-Python-Client==2.8.1
click==8.1.3
pytest==7.2.1
PyYAML==6.0
